+++ src/type/isNullish.ts
@@ -5,8 +5,15 @@ import { isUndefined } from "./isUndefined.js";
 /**
  * Check if `entity` is `undefined` or `null`.
  *
+ * @example
+ * ```typescript
+ * isNullish(undefined); // true
+ * isNullish(null); // true
+ * isNullish(1); // false
+ * ```
  * @template Actual Actual type of the entity.
  * @param entity Entity to check.
+ * @returns `true` if is nullish, `false` if not.
  */
 export const isNullish = <Actual = unknown>(
 	entity: Actual | Nullish,