+++ src/types/DecimalTuple.ts
@@ -1,4 +1,4 @@
 /**
- * 2-Tuple to represent decimals [coefficient, exponent]
+ * 2-Tuple to represent decimals [coefficient, exponent].
  */
 export type DecimalTuple = readonly [coefficient: number, exponent: number];