+++ src/types/Grouped.ts
@@ -1,13 +1,14 @@
 import type { ImmutableRecord } from "./ImmutableRecord.js";
 
 /**
- * Object with the following shape:
- * ```
+ * Group object.
+ *
+ * @example
+ * ```typescript
  * {
  * 	[groupName]: [...valuesOfGroupName]
  * }
  * ```
- *
  * @template Item Type of the items in the source array.
  */
 export type Grouped<Item> = ImmutableRecord<ReadonlyArray<Item>, string>;