+++ tests/array/arraySliceFrom.test.ts
@@ -1,13 +1,12 @@
-import { suite } from "@vangware/test";
+import type { Test } from "@vangware/test";
+import type { ReadOnlyArray } from "@vangware/types";
 import { arraySliceFrom } from "../../src/array/arraySliceFrom.js";
 
 const array = [0, 1, 2, 3];
 
-export default suite([
-	{
-		given: "an array of numbers",
-		must: "return a slice of it",
-		received: arraySliceFrom(1)(array),
-		wanted: [1, 2, 3],
-	},
-]);
+export default {
+	given: "an array of numbers",
+	must: "return a slice of it",
+	received: arraySliceFrom(1)(array),
+	wanted: [1, 2, 3],
+} as Test<ReadOnlyArray<number>>;