+++ tests/string/stringParseHexadecimal.test.ts
@@ -1,7 +1,7 @@
-import { suite } from "@vangware/test";
+import type { Tests } from "@vangware/test";
 import { stringParseHexadecimal } from "../../src/string/stringParseHexadecimal.js";
 
-export default suite([
+export default [
 	{
 		given: "a string with value '10'",
 		must: "return the number 16 (0x10)",
@@ -14,4 +14,4 @@ export default suite([
 		received: stringParseHexadecimal("0x10"),
 		wanted: 0x10,
 	},
-]);
+] as Tests<number>;