+++ tests/type/isObject.test.ts
@@ -1,10 +1,10 @@
-import { suite } from "@vangware/test";
+import type { Tests } from "@vangware/test";
 import { isObject } from "../../src/type/isObject.js";
 
 const wantedFalse = { must: "return false", wanted: false };
 const wantedTrue = { must: "return true", wanted: true };
 
-export default suite([
+export default [
 	{
 		given: "a BigInt",
 		received: isObject(BigInt(1)) && isObject(BigInt("1")),
@@ -74,4 +74,4 @@ export default suite([
 		received: isObject(undefined),
 		...wantedFalse,
 	},
-]);
+] as Tests<boolean>;