+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 -   This project adheres to [Semantic Versioning][semver].
 -   This project uses [Gitmoji][gitmoji]
 
+## 8.0.4
+
+-   :arrow_up: dependency upgrade.
+-   :white_check_mark: update tests for new `@vangware/test`.
+-   :wrench: add new `.npmrc`.
+
 ## 8.0.3
 
 -   :arrow_up: dependency upgrade.